Update the selected recipient page to show the page title caption (#2367)
* Update the selected recipient page to show the page title caption

https://eaflood.atlassian.net/browse/WATER-5236

We have missed adding the caption to the selected recipients page.

This change adds the page caption to the selected recipients page.

We have updated our base layout to provide the error summary and backlink. This change has updated the content block to use this improved layout.

This change also uses the common validator.

We can not use the layouts 'pageHeadingHtml' as the page title as this does not support the hint text. So we have added the caption manually.