Update flash notification to support the updated layout (#2364)
https://eaflood.atlassian.net/browse/WATER-5234

We have updated the layout to include the error summary and backlink.

We can also provide the notification banner.

This change adds the notification banner to the updated layout, and updates the existing notifications  to align with the improved notification helper. 

This should ensure all notifications follow the same pattern

We are updating the way we use the notifications by having the layout automatically provide the GDS Notification banner macro.

We used to do this in the views: 
```html
{{ govukNotificationBanner({
     titleText: notification.title,
     text: notification.text
}) }}
```

Now we provide the notification to the macro:
```
{% if notification %}
  {{ govukNotificationBanner(notification) }}
{% endif %}
```

In time we will migrate the existing pages to use the improved 'pageContent'.