Update notice setup recipient name to show name and error (#2355)
* Update notice setup recipient name to show name and error

When sending a paper form, we want to allow the user to send the form to an additional recipient.

Our existing page offers a choice between email and post (a letter). We do not want to show the email option to the user.

The simplest way we can do this is by adding a standalone page to enter the recipient name. This will redirect to the generic address service the same as the existing screen.

This change updates the previously scaffolded code to function as expected.

The view will render the name if filled in and error accordingly.

The submit will redirect to the existing generic address journey. It sets the backlink so the journey knows where to come back to.