Fix missing content in abs alert notifications (#2369)
https://eaflood.atlassian.net/browse/WATER-5232

We've spotted that the last alert details in the monitoring station page do not align with the ones shown on the view licence monitoring stations page in some cases.

In [Fix last alert info in view lic. mon. station page](https://github.com/DEFRA/water-abstraction-system/pull/2354), we cover what the problem is, and implement a fix.

We realised that the core issue is the fact that the `licence_monitoring_stations` table is not linked to the `notifications` table. So, we started work on rectifying this by looking to [Add notification ID to licence monitoring stations](https://github.com/DEFRA/water-abstraction-service/pull/2719).

As part of that, we intend to ensure the existing records get populated, as well as new records. And that is where we found the issues that led to this fix.

A confusion when dealing with the legacy tables and our system views is the difference in names. In the legacy tables, columns were often named one thing, only to be referred to in all other materials as something else. We rectify this in the system views, but there are times we need to stick to the legacy names.

We've spotted that all abstraction alerts created by the legacy code embed `licenceGaugingStationId` in the `notifications.personalisation` JSONB field. But the ones we create use `licenceMonitoringStation`. 😬

Both feature `alertType`, but in the legacy code records, this refers to the licence monitoring station record threshold type ('stop', 'reduce', or 'stop_or_reduce'). In our records, it's the type of alert being sent ('warning', 'reduce', 'stop', and 'resume'). The legacy code stores that in a separate field, `sending_alert_type`, which we're not populating. 🤦

So, this change fixes how we're persisting abstraction alert notifications to realign with what the legacy code was doing.

Knowing this fix is in place means we can then go back to writing migrations to fix the existing records and complete our work to link the two tables.