Link Lic. Monitoring Stations to Notifications
https://eaflood.atlassian.net/browse/WATER-5232

We've spotted that the last alert details in the monitoring station page do not align with the ones shown on the view licence monitoring stations page in some cases.

In [Fix last alert info in view lic. mon. station page](https://github.com/DEFRA/water-abstraction-system/pull/2354), we cover what the problem is, and implement a fix.

We realised that the core issue is the fact that the `licence_monitoring_stations` table is not linked to the `notifications` table. So, we started work on rectifying this by looking to [Add notification ID to licence monitoring stations](https://github.com/DEFRA/water-abstraction-service/pull/2719).

We then realised we were also dependent on another property in the `notifications.personalisation` JSONB field to allow us to link from it back to `licence_monitoring_stations` (the **notification-status** job). So, whilst we were at it we updated `water.scheduled_notification (public.notifications)` to add a `licence_gauging_station_id (licence_monitoring_station_id)` column.

Now those are in place we can

- update our views to incorporate the new fields
- update how we create water abstraction alert notifications to include populating `licence_monitoring_station_id`
- update the **notification-status** job to use the new field to find the right `licence_monitoring_stations` record. Then update it with the relevant `notification_id`.