Replace `request` dependency with `got` (#2314)
https://github.com/DEFRA/water-abstraction-team/issues/145

In our [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) repo the [request](https://github.com/request/request) package is used with the `airbrake` plugin `app/plugins/airbrake.plugin.js` to enable it to work with our proxy server.

However, [request](https://github.com/request/request) is no longer maintained and vunerabilities have now crept in that will no longer be fixed. As a result we now need to find a replacement for the deprecated package.

We couldn't find any suitable drop-in replacements for `request`; for example, `@root/request`](https://www.npmjs.com/package/@root/request) and [`requestretry`](https://www.npmjs.com/package/requestretry) looked like they may be suitable but don't appear to be currently maintained, so we could just be swapping one outdated dependency for another.

The solution we devised was to take the `got` package that we currently use, and wrap it in a function that resembles the `request` API:

- It can be called with either a URL string (for a GET request) or an options object.
- It accepts a callback with the signature `(error, response, body)`.
- After making the request, it invokes the callback with either the result or an error.

This results in the drop-in replacement we need.