Link Notifications to Lic. Monitoring Stations (#2374)
https://eaflood.atlassian.net/browse/WATER-5232

We've spotted that the last alert details in the monitoring station page do not align with the ones shown on the view licence monitoring stations page in some cases.

In [Fix last alert info in view lic. mon. station page](https://github.com/DEFRA/water-abstraction-system/pull/2354), we cover what the problem is, and implement a fix.

We realised that the core issue is the fact that the `licence_monitoring_stations` table is not linked to the `notifications` table. So, we started work on rectifying this by looking to [Add notification ID to licence monitoring stations](https://github.com/DEFRA/water-abstraction-service/pull/2719).

We then realised we were also dependent on another property in the `notifications.personalisation` JSONB field to allow us to link from it back to `licence_monitoring_stations` (the **notification-status** job). So, whilst we were at it, we updated `water.scheduled_notification (public.notifications)` to add a `licence_gauging_station_id (licence_monitoring_station_id)` column.

Having added that, `notification_id` in `licence_monitoring_stations` becomes redundant. In fact, it becomes dangerous because it could lead to misinformation. For example, what if I sent a letter first, then an email? Notify will provide me the status update for the email pretty quickly, so we'll flag that as the 'last alert'. But a day later, when we check the letter's status, Notify has updated it to sent.

In this case, we'll update the licence monitoring station to say the letter was the last alert sent, when in fact it was the email.

So, in this change, we focus on sorting the following so we can get back to PR #2354 and fix the main issue.

- update our `notifications` view to incorporate the new field
- update how we create water abstraction alert notifications to include populating `licence_monitoring_station_id`
- update the **notification-status** job to use the new field to find the right `licence_monitoring_stations` record before updating it