Refactor presenter and implement template standard
As we touch a view we are bringing it up to template standard. Hence, `backLink` and `pageTitleCaption` become properties of the page data.

But we were also able to move more of the logic in the view into the presenter itself, simplifying the view and make more of the logic testable.