Update the view to template standard
We update the view to use template standard, which means we can drop components like the back link and the page heading.

We also remove an extraneous div's and attributes.

We make the decision to move to two key buttons together into a button group. This also means we can use the caption property of the table rather than trying to roll our own, which means another way we can remove anything unneeded.