Rename FetchMonitoringStationService
As its fetching both the station and the licence monitoring stations we felt a name change would help indicate that it is doing something more than just fetching the monitoring station.

We also intend to refactor the service to return these records separately, so this also indicates you are getting something more than a single object back.