Update scaffold view.njk template to use page content (#2380)
We recently updated the layout.njk to provide the error summary and backlink to all pages.

To keep the existing pages working, we added the 'pageContent' block inside the content block.

The content block still overrides all the content as before. But when you use the 'pageContent' block, we provide the error summary and backlink.

This change updates the view.njk to use the 'pageContent' block. This means we can remove the error summary and backlink macros from the template.

This change also updates the presenter to provide the backlink and page title as these are common on every page the scaffold is designed to be used for.