Update scaffold submit service to use 'formatValidationResult'
We recently updated the layout.njk to provide the error summary and backlink to all pages.

To keep the existing pages working, we added the 'pageContent' block inside the content block.

With the error summary provided, we needed to establish a new error-handling mechanism that can be resued across any page.

We added the 'formatValidationResult' in the base presenter. This converts a Joi result into the error required for the error summary component.

This change updates the submit service template to use this improved pattern.

This required adding another variable to the scaffold script which is consistent with the rest of the scaffolding env vars.