Use actual latest notification in view monitoring (#2378)
https://eaflood.atlassian.net/browse/WATER-5232

We've spotted that the last alert details in the monitoring station page do not align with the ones shown on the view licence monitoring stations page in some cases.

In [Fix last alert info in view lic. mon. station page](https://github.com/DEFRA/water-abstraction-system/pull/2354), we cover what the problem is, and implement a fix.

We realised that the core issue is the fact that the `licence_monitoring_stations` table is not linked to the `notifications` table properly. We've rectified that in a series of changes.

> When referring to tables and columns, we are using the new system view names rather than the legacy ones

- We added `last_monitoring_station_id` as a field to `notifications`
- We added to the migration a script to populate it for existing abstraction alert notifications
- We updated the notices setup job to populate the field when creating the `notifications` for abstraction alerts
- We updated the **notification-status** job to use the new field when finding which licence monitoring station to update

Now we have this link, we can provide a more accurate result to users when displaying the last alert for a licence. For example, what happens if we send a letter first, then an email? Notify will provide us the status update for the email pretty quickly, so we'll flag that as the 'last alert'. But a day later, when we check the letter's status, Notify has now updated it to sent.

In this case, we'll update the licence monitoring station to say the letter was the last alert sent, when in fact it was the email.

Because the tables are now linked, we can query for the latest notification by date, irrespective of when the status job updated it. We intend to do that in PR #2354, but this might put us back where we started; the view monitoring station page being out of sync with the view licence monitoring station page.

So, this change updates the view licence monitoring station page to get its 'last alert' data from the `notifications` table, rather than what happens to be saved in the `licence_monitoring_stations`.