Update logic for determining last alert
In the tests the orignal sort would work. But on our local system we had a licence with 2 alerts

- 9 September 2025
- 7 September 2025

When we viewed this the one for the 7 September was displayed as the latest. If we switched the args in the sort, it then displayed correctly but broke the tests.

Our 'dumb' alternate sort worked for both, hence the switch.