Add back in lost alert detail
When we thought we were going to have just the fields in the licence monitoring station (status and statusUpdatedAt) we had to remove some of the alert detail the 'faulty' logic had.

Now are linking to the latest notification for each licence monitoring station again, we can bring that detail back.