Avoid duplicating alert for single tags
Where a licence has a single tag for a monitoring station, we don't want to duplicate the alert ino on the page. So, if there is only the one tag we just show the top level alert data.

If there are multiple tags, we show the alert data for each.