Fix last alert info in view lic. mon. station page (#2354)
https://eaflood.atlassian.net/browse/WATER-5232

At the end of last year, we developed a replacement for the legacy view monitoring station page (see [WATER-4679](https://eaflood.atlassian.net/browse/WATER-4679)) as part of our ongoing process of dealing with legacy technical debt.

Both the legacy and new pages require us to show the date and type of the last alert sent for each 'tagged' licence to the monitoring station.

Then, as part of [WATER-4980](https://eaflood.atlassian.net/browse/WATER-4980), we created a new view licence monitoring stations page (for reference, a 'tag' in code and the DB is referred to as a 'Licence Monitoring Station (LMS)'). When you click the new 'View' link against a tag, you are taken to a page that displays all the LMS records for that licence and monitoring station.

This would allow users, for the first time, to see all the details related to an LMS, whilst enabling us to simplify the process of removing a 'tag'.

The new page also displays the date of the last alert for the licence across its LMS records.

We've spotted that the last alert details in the monitoring station page do not align with the ones shown on the view licence monitoring stations page in some cases.

When a discrepancy occurs, the LMS view page will either display the details for an older alert or report that no alert has been found.

The issue is two-fold. The first is that the view LMS page obtains its information differently. It is querying the notifications for the last 'water abstraction alert' for the given licence. The query relies on the notification's personalisation containing a `licenceId` property.

Abstraction alerts created by the legacy code have this; those by the new system code (which is any since we went live in August) don't. This explains why, if an older alert is present, it will display the alert details. And if that is genuinely the last alert, the details will be in sync with the view monitoring station page.

If a newer alert has been issued, this one won't be found. Hence, the view LMS page becomes out of sync.

We suspect the difference in how the data is obtained is because of a mistaken assumption about the alert details on the view monitoring page. In the legacy version of abstraction alerts, at the point the notification was sent to GOV.UK Notify it would update the LMS record with the alert type and timestamp. But this didn't account for failed notifications.

Knowing this, the view LMS page instead looks for the last 'sent' water abstraction notification for the licence. This would provide an accurate reflection of when the last alert was, although it might be confusing when the pages show different data.

In fact, the new abstraction alert feature handles this. When the alert is sent, we don't update the LMS record. Instead, we wait for our new nightly 'notification-status' job to check in with Notify on the status of the notification. If Notify confirms it has been sent, we then update the LMS record.

So, as we advance, what we show in the view monitoring station will be accurate, and the view LMS page should obtain the last alert data in the same way.

To ensure data consistency and provide a better link between the notification and the licence, we will make the following two data fixes.

- Update the code that creates the notification record to persist the licence ID as the legacy code did (this change).
- Create a one-off migration script to add the `licenceId to those water abstraction alert records that don't have it (see [Add missing licenceId to abs alert notifications](https://github.com/DEFRA/water-abstraction-service/pull/2717)).

We also update the LMS page to obtain alert details in the same way as the view Monitoring station page. Because a licence can have multiple tags against a monitoring station, and therefore different last alert data, we show the latest at the top of the page, but add a **Last alert sent** line to each tag's detail block.

[WATER-4679]: https://eaflood.atlassian.net/browse/WATER-4679?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[WATER-4980]: https://eaflood.atlassian.net/browse/WATER-4980?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ