Update service
And admit to being naughty!

We had forgotten the test for this was being skipped because of how flakey it is. This being a live issue we don't have the time now to

- resolve the flakey issues
- then update to cover our changes

We've manually tested and all is well. We'll just have to carry the risk and pinky-promise to fix in the near future.