Fix status in view notices page for pending (#2387)
https://eaflood.atlassian.net/browse/WATER-5244

An issue has been spotted with the view notices page.

When viewing the Notices, the status shown is either `SENT` or `ERROR`. However, if you click on a recently created notice, the status against each notification will be `PENDING`.

This is confusing users.

It's simply because the logic behind the view notices page is too simplistic. It is only checking if a notice has an error flag. If it does, it is showing `ERROR`, else `SENT`.

We need to amend the query that is getting the notices to check the linked notifications. We'll then apply a status to the notice on a priority basis.

- If any notifications have a status of `ERROR`, show `ERROR`
- Else if any notifications have a status of `PENDING,` show `PENDING`
- Else show `SENT`

We have a lot of notices, and updating the query to look at `water.scheduled_notification` is going to affect performance. So, to help with that, we are also [Adding some indexes to notifications](https://github.com/DEFRA/water-abstraction-service/pull/2721).