Remove Snyk (#2389)
https://github.com/DEFRA/water-abstraction-team/issues/148

Back in the early days of digital services [Snyk](https://snyk.io/) was one of the first tools we hooked up to our projects. It would periodically scan the projects for vulnerabilities with your dependencies. But it could also be a part of your CI: checking that new dependencies you might add did not have known issues.

Recently, we hit an issue with that CI task because SSO was enabled for the Defra organisation. We initially [resolved the issue](https://github.com/DEFRA/water-abstraction-team/issues/144) by switching to running `snyk test` via an action from GitHub itself. But then we found out we were limited to 200 free tests month.

We were going to look at only running `snyk test` when a PR is merged into `main` but then Snyk started to fail PR's for vulnerabilities we knew about and had already created `.snyk` files for.

Tools like this need to be frictionless, especially when the benefit is slight and we are still protected by other means. So, we've decided to remove **Snyk** rather than persevere with it.