Remove new line characters from headings (#2391)
https://eaflood.atlassian.net/browse/WATER-5200

As part of our work to standardise the view templates we've started using a heading macro. This Heading had additional spacing to make it easier to read and look nice for the devs however this has led to additional new line characters actually being entered on the page. This hasn't impacted the look as they're ignored on the browser but our acceptance tests have failed as they're comparing the entire text string of the element.

This PR updates the macro to remove the new lines/carriage returns.