Update adhoc returns period due date
https://eaflood.atlassian.net/browse/WATER-4977

We need to update the adhoc journey to set a different due date.

The adhoc journey template does not require a start and end date for a return. We need to remove this logic and introduce a new due date which will be set 28/29 days in the future from the time it has been sent.

This change removes the previous setting of the "returnsPeriod" for the adhoc journey.

This change updates the adhoc journey to only set the due date for a "returnsPeriod" or "returnCycle".

This logic is determined by the buisness, we need to set the date to 28 days in the future, unless it is an email then we allow an extra day for processing (29 days).