Fix issue with linking abs. alerts to lic mon stns (#2392)
https://eaflood.atlassian.net/browse/WATER-5232

In testing, we have spotted an issue with the changes we made to link a licence gauging station to a notification when we send an abstraction alert.

In `DetermineLicenceMonitoringStationsService`, we take the result of `FetchMonitoringStationDetailsService` and process it, ready for storing in the session.

The abs alert setup journey wil then use this information rather than retrieving it multiple times during the journey.

The setup journey needs the abstraction start and end days and months. But it doesn't care where it comes from. If a licence monitoring station is linked to a condition, it will come from the linked licence version purpose. If it is not, it will come from the licence gauging station itself.

In `DetermineLicenceMonitoringStationsService` we do this.

```javascript
function _licenceMonitoringStations(licenceMonitoringStations) {
  return licenceMonitoringStations.map((licenceMonitoringStation) => {
    const { licenceVersionPurposeCondition, ...rest } = licenceMonitoringStation

    return {
      ...rest,
      ..._licenceVersionPurpose(licenceVersionPurposeCondition),
      notes: _notes(licenceVersionPurposeCondition),
      thresholdGroup: _thresholdGroup(rest.measureType, rest.thresholdValue, rest.thresholdUnit)
    }
  })
}

function _licenceVersionPurpose(licenceVersionPurposeCondition) {
  if (licenceVersionPurposeCondition?.licenceVersionPurpose) {
    return licenceVersionPurposeCondition.licenceVersionPurpose
  } else {
    return {}
  }
}
```

Before the change, we were not fetching `ID` for the `licenceVersionPurpose`. We now include that, but this means when a licence gauging station is linked via a condition to a `licenceVersionPurpose`, the purpose's ID is overwriting the licence monitoring station's ID because of the order of the spread operators.

This change fixes the issue.