Update adhoc licence page validation (#2386)
* Update adhoc licence page validation

https://eaflood.atlassian.net/browse/WATER-5245

We found a bug in our validation logic for a licence (could be more). When validating the licence we where not checking 'AND rl.metadata->>'isCurrent' = 'true' for the return log.

We do check this in the recipients. So the user was getting through the journey and seeing 0 recipients.

This change adds 'AND rl.metadata->>'isCurrent' = 'true' to the fetch used for validation a licence for the adhoc journey.

As we are 'touching' this part of the code, we have updated the view to template standard, used the common validator and updated the test to suit.

This change also uses the 'generateLicenceRef' helper to ensure the licence ref we are using in they are representative of the real licence ref.