Update the check page to show no recipients message and remove the backlink (#2388)
* Update the check page to show no recipients message and remove the backlink

https://eaflood.atlassian.net/browse/WATER-5246

We found a 'bug' when there are no return logs due for the selected return period. We were still showing the regular page with buttons and an empty list.

This change updates the check page to show a message when no recipients are found.

This created an issue when going back (using the backlink) and selecting a different return period. The selected recipients were already set. We do not have a backlink on any check pages.

We have (finally) removed the backlink completely from this page.