Add logic to create return forms notifications (#2398)
* Add logic to create return forms notifications

https://eaflood.atlassian.net/browse/WATER-5226

We have previously added the ability to preview the return forms. When we implemented this, we added the service cal to gutenberg and returned the file (buffer) to the preview service to display.

We are now ready to start sending the PDF file to notify.

To send the PDF file, we need to generate the file for every recipient (this could be many), we need to do this per due return log (that the user has selected).

So we first need to iterate over the selected due return logs and then the recipients to create the notification data.

This change implements this logic only.

We will add the specific notification data in a later change, in the form of a presenter.

This will allow us to use our existing batch notification service.