Clean up notice logic before adding return forms
https://eaflood.atlassian.net/browse/WATER-5226

We have spotted a few issues when developing the return forms notice journey.

This change updates the Notify update presenter to all the body content to be empty. This is because the return forms letter does not return any content.

This change also adds the return log id for the 'FetchReturnsDueByLicenceRefService'. This id will be used in the saving of the notification.

This change also removes some straggling 'false' flags we previously used in 'FetchRecipientsService'. This is no longer needed.