Add start dates for versions in history page
https://eaflood.atlassian.net/browse/WATER-5259

When we were working on updating the service to take over returns management from NALD, there was a concern that the lack of access to modlog data would be an impediment.

To alleviate the concern and to ensure it didn't become a distraction from progressing the returns work, we put together a licence history page.

This combined licence, charge and return version data with the modlogs imported from NALD and displayed them in a table in descending date created order.

If the need for it became critical, the first draft would have been iterated, and we'd have decided from where users can access it.

As we suspected, not being able to see the modlog was not a blocker to WRLS taking over returns management, so it has never become an official feature of the service. Only those who know the URL can find it!

Now, we are working on [View my licence 2.0](https://eaflood.atlassian.net/browse/WATER-5070). This is a redesign of the licence summary page that will incorporate an alternate licence history page.

Again, the intent is to combine version and modlog information. The challenges with the NALD modlog are

The number of records and understand which ones are meaningful

The lack of data in a lot of the records

We're not saying what was done in the draft is right, but it might help inform the work that is going on for **View my licence 2.0**.

The one thing the team have said would help is if we displayed the start date for the versions. They could then compare what we're showing to what users can see in NALD, because they'll be better able to match records. If we've nailed it, it will save them a bunch of analysis work.

So, this change adds the start date for each version to the our existing view history page.