Add return forms to batch notifications (#2400)
* Add return forms to batch notifications

https://eaflood.atlassian.net/browse/WATER-5226

We have implemented a service to generate the PDF file for the return forms ('DetermineReturnFormsService'). This service needs to generate the PDF file and the notification data (through a presenter).

This is different to our existing pattern where we just use the presenters to return the notification. This is necessary to generate the PDF file.

Other than this deviation, we can add the return forms to the batch notifications following the existing pattern.

This change adds the logic to send and save a return forms notification.

This change also updates the 'DetermineReturnFormsService' to return the expected data in the notification table. This aligns with our existing notification presenters. This will get its own presenter in another change.