Set return logs `updatedAt` when received or submitted (#2401)
https://eaflood.atlassian.net/browse/WATER-5256

It has been noticed that a return logs `updatedAt` field is not being updated when a return log is recorded as being received or submitted.

This PR will fix this issue by ensuring that the `updatedAt` field gets updated when a return log is recorded as received or submitted.

No work was required to fix the `updatedAt` field when the return log was recorded as being received as this was actually working (see line 52 in `app/services/return-logs/setup/submit-submission.service.js`). It just wasn't working when the return was submitted.