Fix WAA being sent to old/removed contacts
https://eaflood.atlassian.net/browse/WATER-5260

When sending abstraction alerts we were still sending the alerts to previous licence owners.

We missed some logic where we needed to check the licence document role was the current role.

This change updates the abstraction alerts fetch recipients to check if the licence document role is the 'current'. This is done by the end date being set to a date in the future or the edn date bing null.