Fix WAA being sent to old/removed contacts (#2411)
https://eaflood.atlassian.net/browse/WATER-5260

When sending abstraction alerts we were still sending the alerts to previous licence owners.

We missed some logic where we needed to check the licence document role was the current role.

---

So, we thought customer contacts, being managed in WRLS, were their own thing linked directly to the licence.

Our logic fetches all customer contacts flagged for abstraction alerts for the licence.

We now understand that they are linked to the ‘current’ licence version by way of the company.

For example, we have a new licence and therefore one licence version. That version’s licence holder is ACME Ltd.

This gets imported to WRLS in various places. The critical one is `crm_v2.document_roles` (it was always going to be CRM messing things up!)

|Document role Id|Licence|Role          |Company   |Start date|End date  |
|----------------|-------|--------------|----------|----------|----------|
|12345           |01/123| Licence holder|ACME Ltd  |2025-04-01|          |

When you add a customer contact in WRLS, a `crm_v2.company_contacts` record is created (I’ve simplified things for the example!)

|Company contact Id|Company |Email                  |
|------------------|--------|-----------------------|
|98765             |ACME Ltd|uma.leggerts@acme.co.uk|

Now, let's say you issue a new licence version in NALD, because something has changed on the licence. This will result in a new `crm_v2.document_roles` record from the overnight import. An end date will also be applied to the existing one.

|Document role Id|Licence|Role          |Company   |Start date|End date  |
|----------------|-------|--------------|----------|----------|----------|
|12345           |01/123 |Licence holder|ACME Ltd  |2025-04-01|2025-05-31|
|54321           |01/123 |Licence holder|ACME Ltd  |2025-06-01|          |

You don’t have to do anything with customer contacts. They will continue to show because they are linked to ACME Ltd, not the licence! And because the ‘current’ licence holder is still connected to ACME Ltd, they and Uma Leggerts will show in the contact details tab.

Now, the licence is to be transferred to a new licence holder. This will result in a new version being issued in NALD, so a new `crm_v2.document_roles` record.

|Document role Id|Licence|Role          |Company   |Start date|End date  |
|----------------|-------|--------------|----------|----------|----------|
|12345           |01/123 |Licence holder|ACME Ltd  |2025-04-01|2025-05-31|
|54321           |01/123 |Licence holder|ACME Ltd  |2025-06-01|2025-07-31|
|876123          |01/123 |Licence holder|FOOBAR Inc|2025-08-01|          |

This means when you view the licence’s contact details, you will see FOOBAR Inc only. Uma Leggerts will no longer be shown as a customer contact.

---

This change updates the abstraction alerts fetch recipients to check if the licence document role is the 'current'. This is done by the end date being set to a date in the future or the end date being null.