Update return forms to handle no due date (#2409)
https://eaflood.atlassian.net/browse/WATER-4977

We recently implemented logic to handle a return log when no due dates are set. When no due date is set we set it to 28 days in the future (or 29 for a letter).

This change adds this change to the return forms. When a due date is not set we want to set the date to 29 days from today (when the return form is sent)