Add presenter to format the return forms data into a notification (#2406)
* Add presenter to format the return forms data into a notification

https://eaflood.atlassian.net/browse/WATER-5226

We have implemented the logic to send and save a return forms notification. This was already saving the raw page data. But this is not what we needed to save and was a placeholder.

This change adds the presenter for the return forms notification.

A return forms notification differs from our other notification in that is sending a file. This is passed around as an array buffer. This is passed to Notify as 'content' so we map is in the presenter.

The rest of the presenter shapes the data to conform to our existing pattern (and legacy code).

This change has required an update to the fetch to provide the return log id and other region data.

We took this opportunity to introduce a 'dueReturnLog' fixture which aligns to our fetch response. This has allowed us to remove all the due return log objects in our tests.