Fix when `createdBy` added to a return submission (#2405)
https://eaflood.atlassian.net/browse/WATER-5256

Currently when a return is input the `return_submissions.created_by` is only set when a note has been added. This isn't quite right, `created_by` should be getting set whenever a `return_submissions` record is created. This issue is going to be fixed in this PR.

Also in this PR, a change is going to be made to ensure the same timestamp is used for the `return_log.updated_at`, `return_submission.created_at` and `return_submission_lines.created_at` as these shoul be exactly the same. But currently they are a few milliseconds out as these timestamps are generated independently.