Remove the skip and update the tests
We missed this fault with the service because the unit tests were being skipped.

So, we have re-written them in the hope we now have removed their 'flakiness'.

The date from and to filter tests are failing, but this is to be expected as they prove we have an error in our implementation.