Fix error when using date filter in view notices (#2412)
https://eaflood.atlassian.net/browse/WATER-5265

The View Notices page now features a series of filters. They include being able to search by sent date from and to.

In [Fix status in view notices page for pending](https://github.com/DEFRA/water-abstraction-system/pull/2387), we amended the query to ensure the status shown matched the status of the linked notifications.

However, it appears that the change has introduced a bug that means if you attempt to use either Sent date filter, you'll see an error.

This change fixes the query that is causing the error.