OH MY! No messageType in FetchNotifications
In our first commit we removed the status `delivered` from those used by letters as its not a status notification provide according to their docs.

We then made our correction to map 'accepted' to 'pending', updated the presenter tests and assumed all was done.

However, the `ProcessNotificationStatusService` tests started failing.

This surprised us, because the test only sets up an email notification, and mocks Notify responding with 'delivered' for it. We haven't changed anything to do with emails in the presenter, yet it was now returning `null` for the notification's status.

Hmm?

When we dug in we found the issue. The `app/presenters/jobs/notifications/notify-status.presenter.js` determines how to map the Notify status based on the notification's `messageType`. We could see that `test/services/jobs/notification-status/process-notification-status.service.test.js` was setting `messageType`, so why was it not being picked up?

It is because `app/services/jobs/notification-status/fetch-notifications.service.js` was never selecting it! This means since we put the job live, it's interpreted _every_ notification as a letter!