Fix notification-status job for accepted letters (#2415)
https://eaflood.atlassian.net/browse/WATER-5266

We were asked to look into why a water abstraction resume alert that was showing as SENT in view notices was [not appearing on the licence's communications tab](https://eaflood.atlassian.net/browse/WATER-5262).

When we looked, we found the query was looking at the `water.scheduled_notification` table's `notify_status` field, rather than the `status` field, which is what the view notices page uses.

The query was fetching only those where the notify_status was either 'delivered' or 'received'. As far as [GOV.UK Notify](https://docs.notifications.service.gov.uk/rest-api.html#letter-status-descriptions) is concerned:

- **Delivered** means an email was successfully sent
- **Received** means a letter was successfully sent to the printers for sending

So, the query wasn't wrong, just not consistent with how we deem a notification to be successful. In the view notices page, it is when `status` is 'sent'.

In the case of the missing alert

- `notify_status` was 'accepted'
- `status` was 'sent

This is why it was showing in one place, but not in the communications tab.

The notification-status job, when it gets the status from **Notify** is deeming 'accepted' for letters as they have been successful. However, there is still a chance **Notify** can report a 'permanent-failure'.

> The provider cannot print the letter. Your letter will not be dispatched.

We should only be updating the status of letters to 'sent' when Notify reports that their status is 'received'.

---

Whilst making this change, we also spotted that `app/services/jobs/notification-status/fetch-notifications.service.js` was never selecting `messageType`. This gets passed to `app/presenters/jobs/notifications/notify-status.presenter.js`, which uses the value to determine how to map the response from Notify.

This means since we put the job live, it's interpreted _every_ notification as a letter!

We've fixed that now 😁