Housekeeping - use notification model
After working with notifications over the last year, we now know that nothing is ever 'scheduled', so its a completely bogus name for the table.

We added a duplicate PostgreSQL view, and Objection.js model called `NotificationModel`, and we've been doing all our latest work using it, with the intent of dropping `ScheduledNotificationModel` as soon as we can.