Fix missing notifications in licence comms tab (#2416)
https://eaflood.atlassian.net/browse/WATER-5262

We were asked to look into why a water abstraction resume alert that was showing as **SENT** in view notices was [not appearing on the licence's communications tab](https://eaflood.atlassian.net/browse/WATER-5262).

When we looked, we found `FetchCommunicationsService` was looking at the `water.scheduled_notification` table's `notify_status` field, rather than the `status` field, which is what the view notices page uses.

The query was fetching only those where the `notify_status`` was either 'delivered' or 'received'. As far as [GOV.UK Notify](https://docs.notifications.service.gov.uk/rest-api.html#letter-status-descriptions) is concerned:

- **Delivered** means an email was successfully sent
- **Received** means a letter was successfully sent to the printers for sending

So, the query wasn't wrong, just not consistent with how we deem a notification to be successful. In the view notices page, it is when `status` is 'sent'.

[Fix notification-status job for accepted letters](https://github.com/DEFRA/water-abstraction-system/pull/2415) is ensuring we don't set 'accepted' letters to **SENT** going forward. This change updates `FetchCommunicationsService` to use the `status` field instead of `notify_status`, making it consistent with how they are shown in view notices.