Allow ended 'Not due yet' returns to be submitted
https://eaflood.atlassian.net/browse/WATER-5144

The business have needed to revoke a licence, which results in a split return log. Basically, its return period was the full cycle (2025-04-01 to 2026-03-31), which means it's due date was set 28 days after both the period and the cycle end (2026-04-28).

When the licence gets revoked, the return log is replaced with one whose end date matches the licence end date, for example 2025-09-19.

But the return log will still be linked to the same cycle, so the due date will still be calculated as 2026-04-28.

In the service, it will display as **NOT DUE YET**. When we added this status and took over returns management, the decision from the business was that it should not be possible to submit these return logs.

But now they have realised they need to.

This change amends the logic for the action button. Rather than only showing the **Submit** button when we hit 28 days before the due date, we'll make it available as soon as the current date is greater than the return log's end date.