Fix two know flakey unit tests
https://github.com/DEFRA/water-abstraction-team/issues/139

We have some know 'flakey' unit tests. One is the `test/services/notices/fetch-notices.service.test.js`, the other is `test/services/billing-accounts/change-address.service.test.js`.

**fetch-notices.service.test.js**

This test includes assertions that the paginated results from the fetch service includes a test record we know we've populated. When run in isolation it always passes. When run in CI it becomes flakey.

The reason is because there are a number of other tests that populate notices. So, depending on when the test runs (we randomise the run order of our unit tests), if it is towards the end, then there will more than 25 notices (our default pagination value), causing the assertion to fail as it only looks at the initial page of results.

To fix this, we've agreed that when we are testing a fetch service that uses pagination, if asserting results contain something we'll stub `defaultPageSize` to a value large enough that all results are returned.

We'll also add a specific paging test, where we'll stub `defaultPageSize` to 'number of test records - 1', i.e. a known value that will cause the fetch service under test to paginate.

**change-address.service**

This test includes assertions that when an action is performed that results in a change to a record, that this is captured in its `updatedAt` field.

The issue we've found is that on fast machines, or when other tests error (this seems to speed up the tests!), the record's `updatedAt` doesn't appear to change.

It's not because the update hasn't been applied; we know it has. Its just purely that the timestamps remain the same because not enough time has lapsed for the vale to be different.

In some of the tests we set the `updatedAt` when we create the test record, to ensure when we update the record it _will_ be later. But we haven't been consistent.

We'll go back through the tests and ensure we always apply an `updatedAt` when creating a record.